Develop - github action fixes

+++ scripts/build.sh
@@ -153,10 +153,12 @@ fi
 
 if [ "$APP" == "Splunk_CiscoDNACenter" ]; then
   log_debug "Merging local.meta and removing file"
-  # cat "${APP}/metadata/default.meta" "${APP}/metadata/local.meta" > "${APP}/metadata/merged.meta"
-  # mv "${APP}/metadata/merged.meta" "${APP}/metadata/default.meta"
-  rm "${APP}/metadata/local.meta"
-  # chmod 600 "${APP}/metadata/default.meta"
+  if [ -e "${APP}/metadata/local.meta" ]; then
+    # cat "${APP}/metadata/default.meta" "${APP}/metadata/local.meta" > "${APP}/metadata/merged.meta"
+    # mv "${APP}/metadata/merged.meta" "${APP}/metadata/default.meta"
+    rm "${APP}/metadata/local.meta"
+    # chmod 600 "${APP}/metadata/default.meta"
+  fi
 fi
 
 # Re-compress to tar file and clean up
@@ -165,8 +167,10 @@ tar czf "${ROOT}/_build/${FILENAME}" "${APP}"
 # Set variables for GitHub Actions to pick up in subsequent steps
 if [ "$GITHUB_ACTIONS" == "true" ]; then
     log_debug "Setting GitHub Actions outputs"
-    echo ::set-output "name=path::_build/$FILENAME"
-    echo ::set-output "name=file::$FILENAME"
+    # echo ::set-output "name=path::_build/$FILENAME"
+    # echo ::set-output "name=file::$FILENAME"
+    echo "name=path::_build/$FILENAME" >> $GITHUB_OUTPUT
+    echo "name=file::$FILENAME" >> $GITHUB_OUTPUT
 fi
 
 log_debug "path: _build/$FILENAME"
